Fixed Vehicle subclass editing

+++ vehicle/utils.py
@@ -3,12 +3,13 @@
 from vehicle import constants as veh_const, models as veh_models
 
 
-def delete_associated_subclass(listing, vehicle_type):
+def delete_associated_subclass(listing):
     """Delete previously associated subclass of vehicle before changing it"""
-    if vehicle_type == veh_const.SEDAN:
+    if veh_models.Sedan.objects.filter(vehicle=listing.vehicle).exists():
         veh_models.Sedan.objects.filter(vehicle=listing.vehicle).delete()
-    elif vehicle_type == veh_const.TRUCK:
+    elif veh_models.Truck.objects.filter(vehicle=listing.vehicle).exists():
         veh_models.Truck.objects.filter(vehicle=listing.vehicle).delete()
-    elif vehicle_type == veh_const.COUPE:
+    elif veh_models.Coupe.objects.filter(vehicle=listing.vehicle):
         veh_models.Coupe.objects.filter(vehicle=listing.vehicle).delete()
-    veh_models.SUV.objects.filter(vehicle=listing.vehicle).delete()
+    elif veh_models.SUV.objects.filter(vehicle=listing.vehicle).exists():
+        veh_models.SUV.objects.filter(vehicle=listing.vehicle).delete()
